ai-services-fastapi/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
xgboost
openai # For content generation, if using OpenAI API
python-dotenv # For loading environment variables from .env file
joblib # For loading models
